requirements.txt
streamlit
pandas
numpy
plotly
statsmodels
scipy
ruptures
pytest
chronos-forecasting
torch
